+++ user.h
@@ -1,5 +1,4 @@
 #include "stddef.h"
-#include "sys/types.h"
 
 
 struct stat;
@@ -27,7 +26,7 @@ int getpid(void);
 char* sbrk(int);
 int sleep(int);
 int uptime(void);
-void *mmap(void *addr, size_t length, int prot, int flags, int fd, off_t offset);
+void *mmap(void *addr, size_t length, int prot, int flags, int fd, int offset);
 int munmap(void *addr, size_t length);
 
 // ulib.c
